+++ tests/test_export_id.py
@@ -1,4 +1,5 @@
 import json
+import os
 
 try:
     from pathlib import Path
@@ -20,3 +21,9 @@ def test_export_id(app, status, warning):
     assert "FLOW_1" in content_obj['versions']['1.0']['filters'].keys()
     assert "TABLE_1" in content_obj['versions']['1.0']['filters'].keys()
     assert "LIST_1" in content_obj['versions']['1.0']['filters'].keys()
+
+
+@with_app(buildername='html', srcdir='doc_test/doc_export_id')
+def test_export_id_html(app, status, warning):
+    app.build()
+    assert not os.path.exists(os.path.join(app.outdir, 'needs.json'))